+++ csf_assign03/debug.cpp
@@ -1,11 +1,13 @@
 #include "debug.h"
 
+#include <cmath>
+
 #include "cache.h"
 
 std::string i_to_binarystr(uint32_t x, size_t size) {
   std::string bit_str = std::bitset<32>(x).to_string();
   bit_str = bit_str.substr(bit_str.length() - size);
-  for (int i = 4; i < bit_str.length(); i += 5) bit_str.insert(i, " ");
+  for (size_t i = 4; i < bit_str.length(); i += 5) bit_str.insert(i, " ");
   return bit_str;
 }
 
@@ -18,9 +20,9 @@ void debug_print(const Cache &c) {
 
   size_t tag_bits = 32 - log2(c.config.num_sets) - log2(c.config.block_size);
 
-  for (int i = 0; i < c.config.num_sets; i++) {
+  for (uint32_t i = 0; i < c.config.num_sets; i++) {
     bool is_set_any_valid = false;
-    for (int j = 0; j < c.config.num_blocks; j++)
+    for (uint32_t j = 0; j < c.config.num_blocks; j++)
       if (c.sets[i][j].valid) {
         is_set_any_valid = true;
         break;
@@ -29,7 +31,7 @@ void debug_print(const Cache &c) {
 
     std::cout << std::left << std::setw(5) << i;
 
-    for (int j = 0; j < c.config.num_blocks; j++) {
+    for (uint32_t j = 0; j < c.config.num_blocks; j++) {
       std::string tag_str = i_to_binarystr(c.sets[i][j].tag, tag_bits);
 
       if (j != 0) std::cout << std::string(5, ' ');
